requirements.txt
bcrypt
cffi
cryptography
dependency-injector
fastapi
httpx
mypy
mysql-connector-python
pytest
pytest-asyncio
pytest-cov
pytest-dependency
python-dotenv
python-jose
python-multipart
pyyaml
uvicorn